requirements.txt
flask
ollama
python-dotenv
Pillow
